+++ CHANGELOG.md
@@ -10,6 +10,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Technical details
 - Fixed generation of api keys for users
 - Deletes previous pending operator documents
+- Added the app name to the context
+
+### Fixes
+- Observations tool
+  - The monitors observations are now displayed (even if they haven't yet been approved)
 
 ## [0.9.1] - 2017-08-30
 ### Additions